models/sampling/graph_based_sampler.py
"""
Graph-Based Sampling implementation for ProteinFlex.
Implements message passing and local structure preservation.
"""

import torch
import torch.nn as nn
import torch.nn.functional as F
from typing import Dict, Tuple, Optional, List

class MessagePassingLayer(nn.Module):
    def __init__(
        self,
        node_dim: int,
        edge_dim: int,
        hidden_dim: int,
        dropout: float = 0.1
    ):
        """Initialize message passing layer."""
        super().__init__()
        self.node_update = nn.Sequential(
            nn.Linear(node_dim + edge_dim + node_dim, hidden_dim),
            nn.LayerNorm(hidden_dim),
            nn.ReLU(),
            nn.Dropout(dropout),
            nn.Linear(hidden_dim, node_dim)
        )

        self.edge_update = nn.Sequential(
            nn.Linear(edge_dim + 2 * node_dim, hidden_dim),
            nn.LayerNorm(hidden_dim),
            nn.ReLU(),
            nn.Dropout(dropout),
            nn.Linear(hidden_dim, edge_dim)
        )

    def forward(
        self,
        nodes: torch.Tensor,
        edges: torch.Tensor,
        adjacency: torch.Tensor
    ) -> Tuple[torch.Tensor, torch.Tensor]:
        """
        Forward pass for message passing.

        Args:
            nodes: Node features [batch_size, num_nodes, node_dim]
            edges: Edge features [batch_size, num_nodes, num_nodes, edge_dim]
            adjacency: Adjacency matrix [batch_size, num_nodes, num_nodes]

        Returns:
            Updated node and edge features
        """
        B, N, D = nodes.shape

        # Aggregate messages for nodes
        messages = torch.einsum('bijk,bik->bij', edges * adjacency.unsqueeze(-1), nodes)
        node_input = torch.cat([nodes, messages], dim=-1)
        nodes_updated = nodes + self.node_update(node_input)

        # Update edges
        node_pairs = torch.cat([
            nodes.unsqueeze(2).expand(-1, -1, N, -1),
            nodes.unsqueeze(1).expand(-1, N, -1, -1)
        ], dim=-1)
        edge_input = torch.cat([edges, node_pairs], dim=-1)
        edges_updated = edges + self.edge_update(edge_input)

        return nodes_updated, edges_updated

class GraphBasedSampler(nn.Module):
    def __init__(
        self,
        node_dim: int = 768,
        edge_dim: int = 64,
        hidden_dim: int = 512,
        num_layers: int = 6,
        dropout: float = 0.1
    ):
        """
        Initialize Graph-Based Sampler.

        Args:
            node_dim: Dimension of node features
            edge_dim: Dimension of edge features
            hidden_dim: Hidden dimension
            num_layers: Number of message passing layers
            dropout: Dropout rate
        """
        super().__init__()
        self.node_dim = node_dim
        self.edge_dim = edge_dim

        # Edge feature initialization
        self.edge_embedding = nn.Sequential(
            nn.Linear(1, edge_dim),  # Distance-based initialization
            nn.LayerNorm(edge_dim),
            nn.ReLU()
        )

        # Message passing layers
        self.layers = nn.ModuleList([
            MessagePassingLayer(node_dim, edge_dim, hidden_dim, dropout)
            for _ in range(num_layers)
        ])

        # Structure preservation network
        self.structure_preserving = nn.Sequential(
            nn.Linear(node_dim, hidden_dim),
            nn.LayerNorm(hidden_dim),
            nn.ReLU(),
            nn.Linear(hidden_dim, 3)  # 3D coordinates
        )

        # Final node update
        self.node_output = nn.Sequential(
            nn.Linear(node_dim, hidden_dim),
            nn.LayerNorm(hidden_dim),
            nn.ReLU(),
            nn.Linear(hidden_dim, node_dim)
        )

    def initialize_edges(
        self,
        coords: torch.Tensor
    ) -> Tuple[torch.Tensor, torch.Tensor]:
        """Initialize edge features from coordinates."""
        B, N = coords.shape[:2]

        # Compute pairwise distances
        diff = coords.unsqueeze(2) - coords.unsqueeze(1)
        distances = torch.norm(diff, dim=-1, keepdim=True)

        # Create adjacency matrix (connect nearby residues)
        adjacency = (distances.squeeze(-1) < 10.0).float()  # 10Å cutoff

        # Initialize edge features
        edges = self.edge_embedding(distances)

        return edges, adjacency

    def forward(
        self,
        nodes: torch.Tensor,
        coords: Optional[torch.Tensor] = None
    ) -> Tuple[torch.Tensor, torch.Tensor]:
        """
        Forward pass for training.

        Args:
            nodes: Node features [batch_size, num_nodes, node_dim]
            coords: Optional 3D coordinates [batch_size, num_nodes, 3]

        Returns:
            Updated node features and predicted coordinates
        """
        if coords is None:
            coords = self.structure_preserving(nodes)

        # Initialize graph structure
        edges, adjacency = self.initialize_edges(coords)

        # Apply message passing layers
        current_nodes = nodes
        current_edges = edges

        for layer in self.layers:
            current_nodes, current_edges = layer(
                current_nodes,
                current_edges,
                adjacency
            )

        # Final node update
        nodes_updated = self.node_output(current_nodes)

        # Predict coordinates
        coords_pred = self.structure_preserving(nodes_updated)

        return nodes_updated, coords_pred

    def sample(
        self,
        batch_size: int,
        seq_len: int,
        device: torch.device,
        temperature: float = 1.0
    ) -> Tuple[torch.Tensor, torch.Tensor]:
        """
        Generate protein features using graph-based sampling.

        Args:
            batch_size: Number of samples to generate
            seq_len: Sequence length
            device: Device to generate on
            temperature: Sampling temperature

        Returns:
            Tuple of (node features, 3D coordinates)
        """
        # Initialize random node features
        nodes = torch.randn(
            batch_size, seq_len, self.node_dim,
            device=device
        ) * temperature

        # Generate initial coordinates
        coords = self.structure_preserving(nodes)

        # Refine through message passing
        nodes_refined, coords_refined = self.forward(nodes, coords)

        return nodes_refined, coords_refined

    def compute_loss(
        self,
        pred_nodes: torch.Tensor,
        target_nodes: torch.Tensor,
        pred_coords: torch.Tensor,
        target_coords: torch.Tensor
    ) -> torch.Tensor:
        """
        Compute training loss.

        Args:
            pred_nodes: Predicted node features
            target_nodes: Target node features
            pred_coords: Predicted coordinates
            target_coords: Target coordinates

        Returns:
            Combined loss value
        """
        # Node feature reconstruction loss
        node_loss = F.mse_loss(pred_nodes, target_nodes)

        # Coordinate prediction loss
        coord_loss = F.mse_loss(pred_coords, target_coords)

        # Distance matrix consistency loss
        pred_dist = torch.cdist(pred_coords, pred_coords)
        target_dist = torch.cdist(target_coords, target_coords)
        dist_loss = F.mse_loss(pred_dist, target_dist)

        return node_loss + coord_loss + dist_loss
